Use coverage stats from `.global_usage`, not `alt-ww`
